Implement processUpload working for v2 without the updates for uploadStatus yet
